requirements.txt
Django==2.0.7
django-admin-rangefilter==0.3.7
djangorestframework==3.7.7
pytz==2018.5
six==1.11.0
